64bit portability bug
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/simgrid/simgrid/trunk@5957 48e7efb5-ca39-0410-a469-dd3cf9ba447f

+++ tools/tesh/tesh.c
@@ -170,7 +170,7 @@ static void parse_environ(){
   for (i=0; environ[i];i++) {
     p=environ[i];
     char *eq = strchr(p,'=');
-    char *key = bprintf("%.*s",eq-p,p);
+    char *key = bprintf("%.*s",(int)(eq-p),p);
     xbt_dict_set(env,key,xbt_strdup(eq+1),xbt_free_f);
     free(key);
   }
